add bmrt.py, separate out from job.py, adjust interfaces
