+++ conbench/app/benchmarks.py
@@ -16,7 +16,7 @@ import conbench.units
 from conbench.app import app
 from conbench.app._endpoint import authorize_or_terminate
 from conbench.config import Config
-from conbench.job import BMRTBenchmarkResult, TBenchmarkName, bmrt_cache
+from conbench.bmrt import BMRTBenchmarkResult, TBenchmarkName, bmrt_cache
 from conbench.outlier import remove_outliers_by_iqrdist
 
 """
