+++ conbench/app/index.py
@@ -9,7 +9,7 @@ from sqlalchemy import select
 
 import conbench.util
 from conbench.dbsession import current_session
-from conbench.job import bmrt_cache
+from conbench.bmrt import bmrt_cache
 
 from ..app import rule
 from ..app._endpoint import AppEndpoint, authorize_or_terminate
